triangle/square nodes & attempt at colored markers
added triangular & square nodes
I think changing arrowhead color based on link color is near impossible. It's an SVG2 feature, which is most likely never going to be released. See this response:
https://stackoverflow.com/questions/16664584/changing-an-svg-markers-color-css
I tried a million svelte workarounds and nothing is working for me. Not worth the time right now but we can revisit